refactor(workstation): extract diff surface + split-diff helpers from inkRuntime (#890 phase 5a.4)
Fourth batch of surface extractions, building on #899 (5a.3). The diff
surface is the largest single surface in inkRuntime — 309 LOC of body
plus 5 supporting split-diff helpers (~110 LOC) that only it consumes.
Lifts both together.

New files:
  - src/workstation/runtime/splitDiff.ts (125 LOC)
      MIN_SPLIT_DIFF_WIDTH, isSplitDiffViable, splitDiffSideProps,
      formatSplitDiffCell, renderSplitDiffBody.
      Lives in runtime/ rather than chrome/ because it composes
      chrome/splitDiff (buildSplitDiffRows) into Ink nodes — it's the
      Ink-rendering wrapper, not the data shaping.
  - src/workstation/surfaces/diff/index.ts (359 LOC)
      renderDiffSurface. Four branches: stash, compare, commit (history
      -> Enter), worktree (status -> Enter). Each respects the
      side-by-side preference and falls back to unified under the
      120-col minimum.

Helpers verified to belong with the diff surface (not the rest of
inkRuntime): post-extraction lint flagged SplitDiffRow, buildSplitDiffRows,
formatStashHeaderIdentity, diffLineProps as fully unused inside
inkRuntime.ts, confirming the diff surface was the sole consumer.

inkRuntime.ts: 4,722 -> 4,305 LOC (-417 this PR; -1,734 cumulative since
5a started, -29% off the original 6,039). 6 imports pruned that became
unused:
  SplitDiffRow, buildSplitDiffRows (chrome/splitDiff)
  formatStashHeaderIdentity (chrome/stashHeader)
  diffLineProps (runtime/utils)
  isSplitDiffViable, renderSplitDiffBody (the new runtime/splitDiff)

11 of 12 main surfaces are now in their own modules. 5a.5 covers the
last main surface (history) + the chrome render helpers (header,
sidebar, footer, overlays, palette); 5a.6 covers the detail / inspector
family. After both, inkRuntime.ts should land around 2,000-2,500 LOC
— mostly the LogInkApp component itself, ready for phase 5b.

Validation: tsc clean, lint clean, 680/680 suites & 6571/6571 tests
passing (same baseline as post-5a.3). No behavior change — diff renders
byte-identically before/after.
